+++ src/components/HomeParts/HomeAnimation.tsx
@@ -10,7 +10,7 @@ const HomeAnimation = () => {
     <Wrapper>
       <div>
         <StyledSpline scene="https://prod.spline.design/LkyNHqmv6VPdWnml/scene.splinecode" onLoad={onLoad} />
-        <StyledSpline scene="https://prod.spline.design/TrfVmPw3GkShfJMY/scene.splinecode" onLoad={onLoad} dark />
+        <StyledSpline scene="https://prod.spline.design/TrfVmPw3GkShfJMY/scene.splinecode" onLoad={onLoad} />
       </div>
     </Wrapper>
   )
