+++ src/contracts/index.ts
@@ -1,3 +1,1 @@
-export { default as GrantsAbi } from './GrantsAbi'
-export { default as ArtizenArtifactsAbi } from './ArtizenArtifactsAbi'
 export { default as SeasonsAbi } from './SeasonsAbi'
