+++ src/gql/artifacts.ts
@@ -2,40 +2,23 @@ import { gql } from '@apollo/client'
 import { ARTIFACT } from './fragments'
 
 export const INSERT_ARTIFACTS = gql`
+  ${ARTIFACT}
   mutation insert_Artifacts($objects: [Artifacts_insert_input!]!) {
     insert_Artifacts(objects: $objects) {
       returning {
-        id
+        ...Artifact
       }
     }
   }
 `
 
 export const UPDATE_ARTIFACTS = gql`
+  ${ARTIFACT}
   mutation update_Artifacts_many($updates: [Artifacts_updates!]!) {
     update_Artifacts_many(updates: $updates) {
       returning {
-        id
+        ...Artifact
       }
-    }
-  }
-`
-
-export const SUBSCRIBE_OPEN_EDITIONS = gql`
-  fragment OpenEditionCopy on OpenEditionCopies {
-    value
-    copies
-    user {
-      id
-      firstName
-      lastName
-      artizenHandle
-      profileImage
-    }
-  }
-  subscription openEditions($where: OpenEditionCopies_bool_exp) {
-    OpenEditionCopies(where: $where) {
-      ...OpenEditionCopy
     }
   }
 `
